+++ rustdoc-json/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
-* Add `Builder::package_target` and `PackageTarget`
+* Add `Builder::package_target()` and `PackageTarget`
 * Correctly determine json path for `Builder::default().package("crate@1.0.0")`
+* Add `Builder::silent()` to suppress stdout and stderr
 
 ## v0.7.3
 * Derive `Clone` for `rustdoc_json::Builder`